fix inspector tests
